[ADD/CHG] 1. 新增Sqlite_orm连接池，解决多线程同时占用同一句柄问题; 2. 改造DAO层接口，使用池化组件；3. 改造其他接口入参

+++ src/work/web/project_vo.cpp
@@ -38,8 +38,8 @@ ProjectVo CovertProjectVo(const Project &p)
         p.m_listenPort,
         p.m_targetIp,
         p.m_userId,
-        p.m_status,
-        p.m_active,
+        static_cast<int32_t>(p.m_status),
+        static_cast<int32_t>(p.m_active),
         p.m_ctime.toString()
     };
 }
